Add_priceProduct_to_orderItem

+++ src/services/orderItems.services.js
@@ -18,13 +18,41 @@ export const getAllOrederItem = () => {
 
 // Creacion nueva
 export const createNewOrderItem = async (data) => {
-    const newOrderItem = await prisma.orderItem.create({
-        data
-    })
-    //calcular el total de la orden despues de agregar el item
-    await calcularTotalAmount(newOrderItem.orderId)
+    try {
+        // Obtener el precio del producto desde la tabla Product
+        const product = await prisma.product.findUnique({
+            where:{id: data.productId}
+        })
 
-    return newOrderItem
+        //verificar si el producto existe
+        if (!product) {
+            throw new Error('Producto no encontrado')
+        }
+
+        //crear el nuevo orderItem con el precio del producto
+        const newOrderItem = await prisma.orderItem.create({
+            data: {
+                orderId: data.orderId,
+                productId: data.productId,
+                quantity: data.quantity,
+                price: product.price //asignar automaticamente el precio del producto
+            }
+        });
+
+        if (!newOrderItem.orderId) {
+            throw new Error('Order Id no encontrado en el OrderItem')
+        }
+
+        //calcular el total de la orden despues de agregar el item
+        await calcularTotalAmount(newOrderItem.orderId)
+
+        return newOrderItem
+
+    } catch (error) {
+        console.error("Error al crear el orderItem",error)
+        throw new Error("Error al crear el item de la orden")
+    }
+    
 }
 
 //Busqueda por Id
@@ -49,20 +77,30 @@ export const getAllOrderItemById = (id) => {
 // }
 //Nuevo eliminacion
 export const deleteOrderItem = async(id) => {
-    //obtener el orderItem antes de eliminarlo, para saber a que orden pertenece
-    const orderItems = await prisma.orderItem.findUnique({
-        where: {id}
-    })
+    try {
+        //obtener el orderItem antes de eliminarlo, para saber a que orden pertenece
+        const orderItems = await prisma.orderItem.findUnique({
+            where: {id}
+        })
+        // Verificar que el OrderItem exista y tenga un orderId válido
+        if (!orderItems) {
+            throw new Error('Item de orden no encontrado');
+        }
 
-    // eliminar el orderItem
-     const deleteOrderItems = await prisma.orderItem.delete({
-         where: {id}
-     })
+        // eliminar el orderItem
+        const deleteOrderItems = await prisma.orderItem.delete({
+            where: {id}
+        })
 
-     // Calcular el total de la orden despues de eliminar el item
-    await calcularTotalAmount(orderItems.orderId)
+        // Calcular el total de la orden despues de eliminar el item
+        await calcularTotalAmount(orderItems.orderId)
 
-    return deleteOrderItems
+        return deleteOrderItems
+    } catch (error) {
+        console.error("Error al eliminar el OrderItem:", error);
+        throw new Error("Error al eliminar el item de la orden");
+    }
+    
 }
 
 //Actualizacion
